requirements.txt
langchain 
openai 
langchain-community 
pymupdf
streamlit
numpy
streamlit-pdf-viewer
python-dotenv